backend/pyproject.toml
[tool.poetry]
name = "thedevstarter-backend"
version = "0.1.0"
description = ""
authors = ["Byteoski Developers OPC Pvt Ltd"]
readme = "README.md"
packages = [{include = "thedevstarter_backend"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django-ninja = "^1.0.1"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
stripe = "^7.6.0"
pyjwt = "^2.8.0"
social-auth-app-django = "^5.4.0"
django-cors-headers = "^4.3.1"
pypdf = "^3.17.1"
openai = "^1.3.7"
uvicorn = "^0.24.0.post1"
sentry-sdk = {extras = ["django"], version = "^1.38.0"}
aiohttp = "^3.9.1"
pydantic = {extras = ["email"], version = "^2.5.2"}
ninja-put-patch-file-upload-middleware = "^0.1.1"
boto3 = "^1.34.2"
django-storages = "^1.14.2"
pillow = "^10.1.0"
django-tinymce = "^3.6.1"
pre-commit = "^3.6.0"
flake8 = "^6.1.0"
django-better-admin-arrayfield = "^1.4.2"
tiktoken = "^0.5.2"
langchain-community = "^0.0.12"
langchain = "^0.1.0"
langchain-openai = "^0.0.2.post1"
qdrant-client = "^1.7.0"
django-async-stream = "^0.3"
nest-asyncio = "^1.6.0"
django = "^5.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
